Photo module improvements
Slight visual revamp
Improved behaviour when setting albums/collections as featured

Improved image uploading flow slightly (Ability to upload EXIF-less files is still missing)

+++ app/routes/api/routes/photos.py
@@ -5,7 +5,9 @@ import os
 import re
 
 import flask
+import sqlalchemy as sa
 from flask_babel import _
+from werkzeug.utils import secure_filename
 
 from app import context, db
 from app.models import Album, Collection, PermissionScope, PermissionType, Photo
@@ -18,19 +20,19 @@ ALBUM_URL = re.compile(r"/photos/(.+?)/(.+?)/?")
 
 
 def _check_rights_and_get_album() -> Album:
-    collection_name = flask.request.form.get("collection")
+    collection_name = flask.request.values.get("collection")
     if not collection_name:
         flask.abort(400, "Form field 'collection' missing")
-    album_name = flask.request.form.get("album")
+    album_name = flask.request.values.get("album")
     if not album_name:
         flask.abort(400, "Form field 'album' missing")
 
     collection = db.session.scalars(db.select(Collection).filter_by(dir_name=collection_name)).first()
     if not collection:
         flask.abort(400, f"Invalid collection: '{collection_name}'")
-    album = next((a for a in collection.albums if a.dir_name == album_name), None)
+    album = db.session.scalars(db.select(Album).filter_by(collection=collection, dir_name=album_name)).first()
     if not album:
-        flask.abort(400, f"Invalid collection: '{collection_name}'")
+        flask.abort(400, f"Invalid album: '{album_name}'")
     if not context.has_permission(PermissionType.write, PermissionScope.album, elem=album):
         flask.abort(403, "Unauthorized")
 
@@ -43,23 +45,31 @@ def upload():
     """Upload a photo to an album (write it to disk and in database)."""
     album = _check_rights_and_get_album()
 
-    file = flask.request.files["file"]
-    if not file:
+    file = flask.request.files.get("file")
+    if not file or not file.filename:
         flask.abort(400, "Missing file (expected as 'file' multipart field)")
 
-    if any(p.file_name == file.filename for p in album.photos):
+    filename = secure_filename(file.filename)
+    if not filename:
+        flask.abort(400, "Invalid filename")
+
+    existing_photo = db.session.scalars(db.select(Photo).filter_by(album=album, file_name=filename)).first()
+    if existing_photo:
         flask.abort(409, "A photo with the same filename already exists in this album.")
 
     # Save photo
-    full_path = os.path.join(album.full_path, file.filename)
+    os.makedirs(album.full_path, exist_ok=True)
+    os.makedirs(album.thumbs_full_path, exist_ok=True)
+    full_path = os.path.join(album.full_path, filename)
     with open(full_path, "wb") as fp:
         fp.write(file.read())
 
     # Register photo
     try:
-        photo = register_new_photo(album, file.filename)
+        photo = register_new_photo(album, filename)
     except PhotoRegistrationError as exc:
-        os.remove(full_path)
+        if os.path.exists(full_path):
+            os.remove(full_path)
         flask.abort(400, " - ".join(exc.args))
 
     db.session.add(photo)
@@ -79,7 +89,7 @@ def edit_photo():
     if not photo_form.validate_on_submit():
         flask.abort(400, "Invalid form body")
 
-    photo = next((p for p in album.photos if p.file_name == photo_form.photo_name.data), None)
+    photo = db.session.scalars(db.select(Photo).filter_by(album=album, file_name=photo_form.photo_name.data)).first()
     if not photo:
         flask.abort(400, "This photo does not exist")
 
@@ -109,13 +119,18 @@ def star_photo():
     if not photo_name:
         flask.abort(400, "Form field 'photo' missing")
 
-    photo = next((p for p in album.photos if p.file_name == photo_name), None)
+    photo = db.session.scalars(db.select(Photo).filter_by(album=album, file_name=photo_name)).first()
     if not photo:
         flask.abort(400, "This photo does not exist")
     if photo.featured:
-        flask.abort(409, "This photo is already the thumbnail of the album")
+        return _("Cette photo est déjà la miniature de l'album."), 200
 
-    photo.album.featured_photo.featured = False
+    db.session.execute(
+        sa.update(Photo)
+        .where(Photo._album_id == album.id)
+        .values(featured=False)
+    )
+    
     photo.featured = True
     db.session.commit()
     return _("Cette photo est maintenant la miniature de l'album."), 200
@@ -124,14 +139,14 @@ def star_photo():
 @sbp.route("/delete_photo", methods=["POST"])
 @context.permission_only(PermissionType.write, PermissionScope.photos)
 def delete_photo():
-    """Make a photo the album thumbnail."""
+    """Delete a photo from the album and disk."""
     album = _check_rights_and_get_album()
 
     photo_name = flask.request.form.get("photo")
     if not photo_name:
         flask.abort(400, "Form field 'photo' missing")
 
-    photo = next((p for p in album.photos if p.file_name == photo_name), None)
+    photo = db.session.scalars(db.select(Photo).filter_by(album=album, file_name=photo_name)).first()
     if not photo:
         flask.abort(400, "This photo does not exist")
